refactor(amazonq): move projectContext folder to amazonq/shared (#4972)
preparation work for #4976
depends on #4970, #4971
