+++ plugins/amazonq/chat/jetbrains-community/resources/META-INF/plugin-chat.xml
@@ -8,7 +8,7 @@
     <projectListeners>
         <listener class="software.aws.toolkits.jetbrains.services.amazonq.toolwindow.AmazonQToolWindowListener"
                   topic="com.intellij.openapi.wm.ex.ToolWindowManagerListener"/>
-        <listener class="software.aws.toolkits.jetbrains.services.cwc.editor.context.project.ProjectContextEditorListener"
+        <listener class="software.aws.toolkits.jetbrains.services.amazonq.project.ProjectContextEditorListener"
                   topic="com.intellij.openapi.fileEditor.FileEditorManagerListener"/>
     </projectListeners>
